Apple cross compile
tested compile on appleArm, ubuntu(arm/intel), pi5. There is a AppleOSX-SDK builder in doc/build-osx.sh this was tested on v26-OSX

+++ Modulator/Modulator.cpp
@@ -304,8 +304,8 @@ qint64 Modulator::readData (char * data, qint64 maxSize)
             float x1=QRandomGenerator::global ()->generateDouble ();
             float x2=QRandomGenerator::global ()->generateDouble ();
 #else
-            float x1=(float)qrand()/RAND_MAX;
-            float x2=(float)qrand()/RAND_MAX;
+            float x1=static_cast<float>(qrand())/static_cast<float>(RAND_MAX);
+            float x2=static_cast<float>(qrand())/static_cast<float>(RAND_MAX);
 #endif
             toneFrequency = m_toneFrequency0 + 0.5*m_fSpread*(x1+x2-1.0);
             m_dphi = m_twoPi * toneFrequency / m_frameRate;
